+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Rhai Release Notes
 Version 1.0.6
 =============
 
+* `MultiInputsStream`, `ParseState`, `TokenIterator`, `IdentifierBuilder` and `AccessMode` are exported under the `internals` feature.
+
 
 Version 1.0.5
 =============